moved throttled_logger to separate package
